requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
amqp==5.0.9
anyio==3.6.1
appnope==0.1.2
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
asttokens==2.0.8
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
billiard==3.6.4.0
bitarray==2.6.0
boto3==1.20.37
botocore==1.23.37
Brotli==1.0.9
buildbot==3.6.1
buildbot-console-view==3.6.1
buildbot-grid-view==3.6.1
buildbot-waterfall-view==3.6.1
buildbot-worker==3.6.1
buildbot-www==3.6.1
cached-property==1.5.2
cachetools==4.2.4
cchardet==2.1.7
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
constantly==15.1.0
construct==2.10.68
construct-typing==0.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect2==0.6.0
cytoolz==0.12.0
decorator==5.1.0
Deprecated==1.2.13
dj-rest-auth==2.2.3
Django==3.2.12
django-admin==2.0.1
django-anymail==8.5
django-cors-headers==3.11.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-extensions==3.1.5
django-filter==21.1
django-oauth-toolkit==1.6.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
django-ses==2.4.0
django-shell-plus==1.1.7
django-six==1.0.4
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
executing==0.10.0
fonttools==4.37.3
forex-python==1.8
frozenlist==1.3.0
future==0.18.2
gevent==22.8.0
greenlet==1.1.3.post0
gunicorn==20.1.0
h11==0.12.0
hexbytes==0.3.0
html5lib==1.1
httpcore==0.15.0
httplib2==0.20.2
httpx==0.23.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.9.0
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
ipdb==0.13.9
ipfshttpclient==0.8.0a2
ipython==7.30.1
isodate==0.6.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.14.0
jwcrypto==1.0
kombu==5.2.3
kri-lib @ git+https://github.com/doyanrebahan/kri_libs.git@005abf5c380569bbb9b8a2772071f3219a89be53
lru-dict==1.1.8
lxml==4.9.1
Mako==1.2.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mongoengine==0.24.2
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
mysql-client==0.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysqlclient==2.1.0
netaddr==0.8.0
numpy==1.22.2
oauth2==1.9.0.post1
oauthlib==3.1.1
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==21.3
pandas==1.4.1
parsimonious==0.8.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
Pillow==9.0.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.24
protobuf==3.19.4
provider==0.1.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.1.2
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.12.0
pydyf==0.3.0
Pygments==2.10.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
pymongo==4.2.0
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyPDF3==1.0.6
pyphen==0.13.0
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.1.3
reportlab==3.6.11
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rlp==2.0.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
screen==1.0.1
sentry-sdk==1.5.6
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
solders==0.9.3
SQLAlchemy==1.4.31
sqlparse==0.4.2
stack-data==0.4.0
supervisor==4.2.4
svglib==1.4.1
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.0
traitlets==5.1.1
Twisted==22.8.0
txaio==22.2.1
types-cachetools==4.2.10
typing_extensions==4.2.0
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.8
uvicorn==0.18.3
uWSGI==2.0.20
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.1
web3==5.30.0
webencodings==0.5.1
wrapt==1.13.3
xhtml2pdf==0.2.8
xlwt==1.3.0
yarl==1.7.2
zeep==4.1.0
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
zopfli==0.2.1
